Added Dependency Injection and UI components

+++ app/src/main/java/com/mecofarid/trending/features/repo/data/source/local/RepoLocalDatasource.kt
@@ -1,5 +1,6 @@
 package com.mecofarid.trending.features.repo.data.source.local
 
+import com.mecofarid.trending.common.data.DataException
 import com.mecofarid.trending.common.data.Datasource
 import com.mecofarid.trending.common.data.Operation
 import com.mecofarid.trending.common.data.Query
@@ -11,14 +12,20 @@ class RepoLocalDatasource(
     private val repoLocalEntityDao: RepoLocalEntityDao
 ): Datasource<List<RepoLocalEntity>> {
     override suspend fun get(query: Query): List<RepoLocalEntity> = when (query) {
-        GetAllTrendingReposQuery -> repoLocalEntityDao.getAllTrendingRepos()
+        GetAllTrendingReposQuery -> getReposOrThrow()
         else -> throw UnsupportedOperationException("Get with query type ($query) is not supported")
     }
 
-    override suspend fun put(query: Query, data: List<RepoLocalEntity>): List<RepoLocalEntity> = when (query) {
-        GetAllTrendingReposQuery -> data.apply {
-            repoLocalEntityDao.deleteAllTrendingReposAndInsert(this)
+    override suspend fun put(query: Query, data: List<RepoLocalEntity>): List<RepoLocalEntity> =
+        when (query) {
+            GetAllTrendingReposQuery -> data.apply {
+                repoLocalEntityDao.deleteAllTrendingReposAndInsert(this)
+            }
+            else -> throw UnsupportedOperationException("Put with query type ($query) is not supported")
         }
-        else -> throw UnsupportedOperationException("Put with query type ($query) is not supported")
+
+    private suspend fun getReposOrThrow() = repoLocalEntityDao.getAllTrendingRepos().also {
+        if (it.isEmpty())
+            throw DataException.DataNotFoundException()
     }
 }
